==> Notebooks/NB1 - Priority Solution_Final.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('../../Data//911_calls_for_service.csv',index_col='callDateTime',parse_dates=True)

data.drop('Unnamed: 0',axis=1,inplace=True)

data.head()

len(data)

data.priority.isnull().value_counts()

grouped = data.groupby(['description','priority']).count()[['callNumber']]

key=grouped.loc[grouped.groupby(level=0).idxmax()['callNumber'].values]

key.head()

grouped.loc[['BURGLARY']]

key.loc[['BURGLARY']]

key.index.is_unique

key = key.reset_index().drop('callNumber',axis=1).set_index('description')['priority'] #key needs to be a Series

key.head()

key.loc['BURGLARY']

key.index.is_unique

mask = data.priority.isnull()

data.loc[mask,'priority'].size

key.loc[data.loc[mask,'description']].count()

key.loc[data.loc[mask,'description']].values

data.loc[mask,'priority'] = key.loc[data.loc[mask,'description']].values

data.priority.isnull().value_counts()

data[data.priority.isnull()]

Unique descriptions with no priority data.<br>
Use similar descriptions to fill missing priorities.

data[data.description.str.contains('CHILD ALONE')]

data.loc[(data.description == 'CHILDREN ALONE') & (data.priority.isnull()),'priority'] = 'Medium'

data[data.priority.isnull()]

data[(data.description.str.contains('CHECK')) & (data.description.str.contains('ADVISE'))].groupby('priority').count()

data.loc[(data.description == 'CK ADVISE') & (data.priority.isnull()),'priority'] = 'Low'

data[data.priority.isnull()]

data.loc[(data.description.str.contains('COURT')) & (data.description.str.contains('VIO'))].groupby('priority').count()

data.loc[(data.description == 'COURT ORDER VIOL') & (data.priority.isnull()),'priority'] = 'Low'

data[data.priority.isnull()]

data[data.description.str.contains('MMEARING')]

data[data.description.str.contains('SMEARING')]

data.loc[(data.description == 'MMEARING') & (data.priority.isnull()),'priority'] = 'Low'

data.loc[data.description == 'MMEARING','description'] = 'SMEARING'

data.priority.isnull().value_counts()

data.to_csv('911_calls_for_service_priority.csv')

data = pd.read_csv('911_calls_for_service_priority.csv')

data.head()

data.index.is_unique

data.priority.isnull().value_counts()
