==> Notebooks/NB2 - Description_Solution_Final.ipynb <==
import pandas as pd
import numpy as np                  

df = pd.read_csv('911_calls_for_service.csv') # Parse the callDatetime column

df.drop('Unnamed: 0',axis=1,inplace=True)

df.head(14)

# no missing data for 'description'
df.isnull().sum()

df['description']=df['description'].str.upper()

df.loc[[140242]]

#Replacing all non-alphanumeric characters with empty string
#df = df.replace({'description': '[\W]'},'',regex=True)
df = df.replace({'description': '[^a-zA-Z0-9/ _]'},'',regex=True)

df.description = df.description.str.replace('/',' ')
df.description = df.description.str.replace('  ',' ')

df.loc[[140242]]

df.description = df.description.str.strip()

df.description[df.description==''].count()

df.loc[df.description=='','description'] = 'NO DESC'

grouped = df.groupby('description').count().reset_index()
grouped.sort_values('priority', ascending=False,inplace=True)
grouped[grouped.priority>1000].count()

grouped[grouped.priority>1000].head(n=10)

grouped[grouped.priority>1000].to_csv('911_calls_descriptions.csv')

df.description.replace({ #FINDING AND REPLACING THE DESCRIPTION
    'INVESTIFATE': 'INVESTIGATE',
    'BAIL-OUT': 'BAILOUT',
    'BALOUT': 'BAILOUT',
    ' BAIL OUT':'BAILOUT',
    'BAILED OUT': 'BAILOUT',
    ' ASSAULT': 'ASSAULT',
    ' ASSAULTED': 'ASSAULT',
    'Y': 'YELLING FOR HELP',
    'YELLING HELP': 'YELLING FOR HELP',
    'YP':'YELLING FOR HELP',
    'ECORT':'ESCORT',
    'WROMG NUMBER': 'WRONG NUMBER',
    'WRONG UNMBER': 'WRONG NUMBER',
    'Wanted on Warr': 'Warrant Service',
    'YELLING FOR HELP':'YELL FOR HELP',
    'YELL 4 HELP': 'YELL FOR HELP',
    'WRECKLESS DRVR': 'WRECKLESSDRIVER',''
    ' BAIL-0UT': 'BAILOUT',
    ' BAILOUT': 'BAILOUT',
    ' BAIL-OUT': 'BAILOUT',
    'WRECKLESS DRVING': 'WRECKLESSDRIVER',
    'WRECKLESS DRV': 'WRECKLESSDRIVER',
    'WRECKLESS DRIVR': 'WRECKLESSDRIVER',
    'WRECKLESS DRIVIN': 'WRECKLESSDRIVER',
    'WRECKLESS DRIVER':  'WRECKLESSDRIVER',
    'WRECKLESS DRIVE': 'WRECKLESSDRIVER',
    'WRECFKLESS DRIVE':'WRECKLESSDRIVER',
    'WRCKLSS DRIVER': 'WRECKLESSDRIVER',
    'WRCKLESS DRIVER': 'WRECKLESSDRIVER',
    ' WRECKLESS DRIVE':'WRECKLESSDRIVER',
    'WRECKLESS': 'WRECKLESSDRIVER',
    'WRECKLES DRIVING': 'WRECKLESSDRIVER',
    'WRECKLES DRIVERS': 'WRECKLESSDRIVER',
    'WRECKLES DRIVER': 'WRECKLESSDRIVER',
    '`INVEST': 'INVESTIGATE',
    '`INVESTIGATE': 'INVESTIGATE',
    'iNVESTIGATE': 'INVESTIGATE',
    '*#TRAFIC ARREST': '*#TRAFFIC ARREST',
    '&75': '& 75',
    '911/No Voice': '911/NO  VOICE',
    '911HANGUP': '911/HANGUP',
    '911 HANGUP': '911/HANGUP',
    'FAMILY DISTURBAN': 'FAMILY DISTURB',
    'FAMILY DISTURBP': 'FAMILY DISTURB',
    'AUTO ACC/INJURY': 'AUTO ACCIDENT',
    '*INVESTIGATE' : 'INVESTIGATE',
    '*TRAFFIC CONTROL'    : 'TRAFFIC CONTROL',
    'CHECK WELLBEING': 'CHECK WELL BEING'}, inplace = True)

df.to_csv('911_calls_for_service_description.csv')

df = pd.read_csv('911_calls_for_service_description.csv')

df.head()
