==> Notebooks/NB3 - Geocoding - Full Address validations.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import time
from pygeocoder import Geocoder

%matplotlib inline

#calls = pd.read_csv('../Project/Data/911_calls_for_service.csv',index_col='callDateTime',parse_dates=True)
calls = pd.read_csv('../Project/Data/911_calls_for_service.csv')
#calls = calls_full.head(n=10)

calls=calls.drop('Unnamed: 0',axis=1)

type(calls.index)

calls.priority.isnull().value_counts()

# Can't analyze repeated calls due to masked numbers.
calls.groupby('callNumber')['callNumber'].count().value_counts()

def set_lat(x):
    try:
        return float(x.split(',')[0][1:])
    except:
        return np.nan
    
def set_lon(x):
    try:
        return float(x.split(',')[1][:-1])
    except:
        return np.nan

calls['lat'] = calls['location'].apply(set_lat)

calls['lon'] = calls['location'].apply(set_lon)

calls[calls.lat.isnull()==True].count()

calls.location = calls.location.replace(to_replace='(,)',value=np.nan)

calls.loc[calls.lat.isnull(),'location'] = calls.loc[calls.lat.isnull(),'incidentLocation']

calls[calls.lat.isnull()==True].head()

calls[(calls.lat.isnull()==True) & (calls.incidentLocation.isnull()==True)]

def fill_coordinates(x):
    try:
        time.sleep(0.01)
        result = Geocoder.geocode(x)
        if (len(result.coordinates) > 0) & (result.city == 'Baltimore'):
            return result.coordinates
        else:
            return np.nan
    except:
        return "ERROR"

results = Geocoder.geocode('ASCIATE')

results.city

type(calls.iloc[0,5])

fill_coordinates(calls.iloc[[0,5]])

fill_coordinates(calls[calls.lat.isnull()==True].iloc[1204,5])

fill_coordinates(calls.loc[calls.lon.isnull()==True,'location'].loc[1204])

calls[calls.lat.isnull()==True].iloc[0,5]

fill_coordinates('3100 POTEE ST')

##### do not run #####
#%timeit
#calls.loc[calls.lon.isnull()==True,'location'] = calls.loc[calls.lon.isnull()==True,'location'].apply(fill_coordinates)
##### do not run #####

calls[calls.lat.isnull()==True].iloc[[0]]

calls.loc[calls.lon.isnull()==True]['location'].value_counts()

pd.Dataf

def find_addr(x):
    if x == '(,)':
        return np.nan
    else:
        lat = float(x.split(',')[0][1:])
        lon = float(x.split(',')[1][:-1])
        time.sleep(1)
        return Geocoder.reverse_geocode(lat,lon).formatted_address

def find_zip(x):
    if x == '(,)':
        return np.nan
    else:
        lat = float(x.split(',')[0][1:])
        lon = float(x.split(',')[1][:-1])
        return Geocoder.reverse_geocode(lat,lon).postal_code

%timeit
calls['address']=calls['location'].transform(find_addr)
calls.to_csv('../Project/Data/911_calls_with_addr.csv')
#OVER_QUERY_LIMIT

calls

calls['zipCode']=calls['location'].transform(find_zip)
#calls.to_csv('../Project/Data/911_calls_with_addrZip.csv')

calls.head()
