==> Notebooks/NB4 - Getting Coordinates using Google Maps URL requests.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import urllib
#import time
#from pygeocoder import Geocoder

%matplotlib inline

#calls = pd.read_csv('../Project/Data/911_calls_for_service.csv',index_col='callDateTime',parse_dates=True)
calls = pd.read_csv('../Project/Data/911_calls_for_service.csv')
#calls = calls_full.head(n=10)

calls=calls.drop('Unnamed: 0',axis=1)

type(calls.index)

calls.head()

calls.priority.isnull().value_counts()

# Can't analyze repeated calls due to masked numbers.
calls.groupby('callNumber')['callNumber'].count().value_counts()

def set_lat(x):
    try:
        return float(x.split(',')[0][1:])
    except:
        return np.nan
    
def set_lon(x):
    try:
        return float(x.split(',')[1][:-1])
    except:
        return np.nan

calls['lat'] = calls['location'].apply(set_lat)

calls['lon'] = calls['location'].apply(set_lon)

calls[calls.lat.isnull()==True].count()

calls.loc[calls.lat.isnull()==True,'location'].value_counts()

calls.location = calls.location.replace(to_replace='(,)',value=np.nan)

#calls.loc[calls.lat.isnull(),'location'] = calls.loc[calls.lat.isnull(),'incidentLocation']

calls.loc[calls.lat.isnull()==True,'location'].size

calls[(calls.lat.isnull()==True) & (calls.incidentLocation.isnull()==True)]

def get_coordinates(addr):
    #addr = '500 N MARTIN LUTHER KING JR BL '

    url = 'https://www.google.ca/maps/place/' + addr.strip().replace(' ','+') + ',+Baltimore'

    response = urllib.request.urlopen(url)

    for x in range(10):
            response.readline()
    data = str(response.readline()).split('],[')[0]
    lat, lon = data.split(',')[1:]
    return '('+lon+','+lat+')'

%%time
addr = '500 N MARTIN LUTHER KING JR BL '
get_coordinates(addr)

calls.loc[calls.lon.isnull()==True].head()

calls.head()

#%%time
#calls.loc[calls.lon.isnull()==True,'location'] = calls.loc[calls.lon.isnull()==True,'location'].apply(get_coordinates)

calls_sub=calls.loc[calls.lon.isnull()==True]

calls_sub.loc[:,'location'].size

calls_sub.iloc[0,6]

#%%time
#for x in range(0,1000):
#    calls_sub.iloc[x,6] = get_coordinates(calls_sub.iloc[x,6])

calls_sub['lat'] = calls_sub['location'].apply(set_lat)

calls_sub['lon'] = calls_sub['location'].apply(set_lon)

calls_sub.iloc[0,6]

# original: 82532
# Round1: 82346

calls_sub.loc[calls_sub.lon.isnull()==True,'location'].size

82532 - 82346
