==> Notebooks/NB6 - Coordinates Consolidation Final.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import time
import datetime
from pygeocoder import Geocoder

records = 82532
data = pd.DataFrame
list(range(0,records,1000))[-1]

records = 82532
for x in range(0,records,1000):    
    floc = './Adil/Adil_coordinates_'+repr(x)+'_'+repr(x+999)+'.csv'
    if Path(floc).is_file():
        patch = pd.read_csv(floc)
        patch = patch.drop(patch.columns[0],axis=1)
        if data.empty:
            data = patch.copy()
        else:
            data = pd.concat([data,patch])
    
    floc = './Mohammad/Mohammad_Ali_coordinates_'+repr(x)+'_'+repr(x+999)+'.csv'
    if Path(floc).is_file():
        patch = pd.read_csv(floc)
        patch = patch.drop(patch.columns[0],axis=1)
        if data.empty:
            data = patch.copy()
        else:
            data = pd.concat([data,patch])
    
    floc = './Asim/Asim_coordinates_'+repr(x)+'_'+repr(x+999)+'.csv'
    if Path(floc).is_file():
        patch = pd.read_csv(floc)
        patch = patch.drop(patch.columns[0],axis=1)
        if data.empty:
            data = patch.copy()
        else:
            data = pd.concat([data,patch])
#Last Patch            
floc = './Adil/Adil_coordinates_'+repr(x)+'_82531.csv'
if Path(floc).is_file():
    patch = pd.read_csv(floc)
    patch = patch.drop(patch.columns[0],axis=1)
    if data.empty:
        data = patch.copy()
    else:
        data = pd.concat([data,patch])

data.info()

print('Success Rate: %.2f' % (data.location[data.location.isnull()==False].size/len(data)*100)
      +'% Total:',data.location[data.location.isnull()==False].size)
print('Missing Rate: %.2f' % (data.location[data.location.isnull()==True].size/len(data)*100)
       +'% Total:',data.location[data.location.isnull()==True].size)

data.callNumber.is_unique

data.reset_index(drop=True, inplace=True);data.head()

calls = pd.read_csv('./911_calls_for_service.csv')

calls.drop('Unnamed: 0',axis=1, inplace=True)

calls.tail()

calls.loc[(calls.location=='(,)') | (calls.location=='(')].count()

data.count()

calls.set_index('callNumber',inplace=True)

calls.index.is_unique

data.set_index('callNumber',inplace=True)

data.index.is_unique

mask = (calls.location=='(,)') | (calls.location=='(')

calls.loc[mask,'location'].count()

data.location.size

data.location.isnull().value_counts()

calls.loc[mask,'location'] = data.location.values

calls.location.isnull().value_counts()

calls.reset_index(inplace=True)

calls.tail()

calls.to_csv('../911_calls_for_service_coordinates.csv')

calls = pd.read_csv('../911_calls_for_service_coordinates.csv')

calls.head()
