==> Notebooks/NB7 - Analysis_Data_Consolidation_Final.ipynb <==
import pandas as pd
import numpy as np

calls = pd.read_csv('911_calls_for_service.csv')
calls.drop('Unnamed: 0',axis=1, inplace=True)
calls_cord = pd.read_csv('911_calls_for_service_coordinates.csv')
calls_cord.drop('Unnamed: 0',axis=1, inplace=True)
calls_pr = pd.read_csv('911_calls_for_service_priority.csv')
calls_pr.drop('Unnamed: 0',axis=1, inplace=True)
calls_desc = pd.read_csv('911_calls_for_service_description.csv')
calls_desc.drop('Unnamed: 0',axis=1, inplace=True)

calls.iloc[[0,-1]]

len(calls)

calls_cord.iloc[[0,-1]]

len(calls_cord)

calls_pr.iloc[[0,-1]]

len(calls_pr)

calls_desc.iloc[[0,-1]]

len(calls_desc)

calls.priority.isnull().sum()

calls.priority = calls_pr.priority.copy()

calls.priority.isnull().sum()

calls.loc[(calls.location == '(,)') | (calls.location == '('),'location'].count()

calls.location = calls_cord.location.copy()

calls.loc[(calls.location == '(,)') | (calls.location == '('),'location'].count()

calls.location.isnull().value_counts()

calls.description = calls_desc.description.copy()

calls[calls.description.str.contains('TEST',case=False)].count()

calls[calls.description.str.contains('TEST',case=False)].head()

calls.drop(calls[calls.description.str.contains('TEST',case=False) == True].index,inplace=True)

calls[calls.description.str.contains('TEST',case=False)].count()

calls.isnull().sum()

len(calls)

calls.set_index('callDateTime',inplace=True)

calls.head()

def set_lat(x):
    try:
        return float(x.split(',')[0][1:])
    except:
        return np.nan
    
def set_lon(x):
    try:
        return float(x.split(',')[1][:-1])
    except:
        return np.nan

calls['lat'] = calls['location'].apply(set_lat)

calls['lon'] = calls['location'].apply(set_lon)

calls.head()

calls.to_csv('911_calls_for_service_clean_final.csv')

calls = pd.read_csv('911_calls_for_service_clean_final.csv',index_col='callDateTime',parse_dates=True)
calls.head()

type(calls.index)
